+++ back-end/src/main/java/com/nightletter/domain/diary/repository/DiaryRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.nightletter.domain.diary.entity.Diary;
 
 @Repository
-public interface DiaryRepository extends JpaRepository<Diary, Integer>, DiaryCustomRepository {
+public interface DiaryRepository extends JpaRepository<Diary, Long>, DiaryCustomRepository {
 
     List<Diary> findDiariesByWriter(Member writer);
 }
